requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy>=2.0
psycopg2-binary
requests
